LayoutTests/fast/forms/ValidityState-valueMissing-001-expected.txt

This test checks validity.valueMissing with blank values, blank options selected, or nothing selected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS valueMissingFor("input") is true
PASS valueMissingFor("textarea") is true
PASS valueMissingFor("select-no-option") is true
PASS valueMissingFor("select-placeholder-selected") is true
PASS valueMissingFor("select-without-placeholder") is false
PASS valueMissingFor("select-placeholder-selected-size2") is false
PASS valueMissingFor("select-without-placeholder-size2") is false
PASS valueMissingFor("select-none-selected-multiple") is true
PASS valueMissingFor("select-fake-placeholder-selected-multiple") is false
PASS valueMissingFor("select-without-fake-placeholder-multiple") is false
PASS valueMissingFor("select-none-selected-size2-multiple") is true
PASS valueMissingFor("select-fake-placeholder-selected-size2-multiple") is false
PASS valueMissingFor("select-without-fake-placeholder-size2-multiple") is false
PASS valueMissingFor("select-optgroup") is false
PASS valueMissingFor("select-disabled-option") is true
PASS successfullyParsed is true

TEST COMPLETE
